+++ src/app/shared/constants/index.ts
@@ -2,3 +2,7 @@ export enum STATUS {
   'Active' = 'active',
   'Completed' = 'completed',
 }
+
+export enum LOCAL_STORAGE_KEY {
+  TodoList = 'to-do-list',
+}
