+++ src/app/shared/models/task.ts
@@ -3,5 +3,5 @@ import { STATUS } from '../constants';
 export interface TaskProps {
   id: string;
   content: string;
-  status: STATUS.Active | STATUS.Completed;
+  status: STATUS;
 }
