+++ src/stylesheet/style.scss
@@ -65,7 +65,7 @@ input[type='checkbox'] {
 
 .home {
   background-image: url('../assets/background.png');
-  min-height: 100vh;
+  height: 100vh;
 
   display: flex;
   align-items: center;
@@ -78,7 +78,6 @@ input[type='checkbox'] {
   padding: 30px;
 
   border-radius: 20px;
-
   background-color: map-get($colors, 'gray-bg');
   box-shadow: 0px 5px 15px rgba(0, 0, 0, 0.1);
 
@@ -148,6 +147,8 @@ input[type='checkbox'] {
   border-bottom: 0.5px solid map-get($colors, 'gray-c');
   &-list {
     margin-top: 20px;
+    max-height: 40vh;
+    overflow-y: scroll;
   }
   &-wrap {
     align-items: center;
